+++ src/routes/sis/programs/router.ts
@@ -3,7 +3,11 @@ import { Router } from "express";
 import { Prisma, UserRole } from "../../../generated/prisma";
 
 import { auth, AuthRequest } from "../../../auth";
-import { programs } from "../../../sis/program/programs";
+import {
+    programs,
+    ProgramRequestData,
+    ProgramBasicRequestData,
+} from "../../../sis";
 
 export const ProgramRouter = Router();
 
@@ -16,17 +20,21 @@ ProgramRouter.post(
         const result = programs.schema.safeParse(data);
 
         if (!result.success)
-            return res.status(400).json({ message: "Invalid program data" });
+            return res.status(400).json({
+                success: false,
+                message: "Invalid program data",
+            } as ProgramRequestData);
 
         const { code, name } = result.data;
         const program = await programs.create(code, name);
 
         return res.status(program ? 200 : 409).json({
+            success: true,
             message: program
                 ? "Program created successfully"
                 : "Program already exists",
             program,
-        });
+        } as ProgramRequestData);
     }
 );
 
@@ -39,7 +47,9 @@ ProgramRouter.put(
         const result = programs.partialSchema.safeParse(data);
 
         if (!result.success)
-            return res.status(400).json({ message: "Invalid program data" });
+            return res.status(400).json({
+                message: "Invalid program data",
+            } as ProgramBasicRequestData);
 
         const success = await programs.update(
             req.params.code as string,
@@ -50,7 +60,7 @@ ProgramRouter.put(
             message: success
                 ? "Updated the program successfully"
                 : "Program does not exist",
-        });
+        } as ProgramBasicRequestData);
     }
 );
 
@@ -61,17 +71,19 @@ ProgramRouter.delete(
     async (req, res) => {
         const { code } = req.params;
         if (!code)
-            return res
-                .status(400)
-                .json({ message: "Invalid or missing program code" });
+            return res.status(400).json({
+                success: false,
+                message: "Invalid or missing program code",
+            } as ProgramRequestData);
 
         const program = await programs.delete(code);
         return res.status(program ? 200 : 404).json({
+            success: true,
             message: program
                 ? "Program deleted successfully"
                 : "Program does not exist",
             program,
-        });
+        } as ProgramRequestData);
     }
 );
 
@@ -82,16 +94,18 @@ ProgramRouter.get(
     async (req, res) => {
         const { code } = req.params;
         if (!code)
-            return res
-                .status(400)
-                .json({ message: "Invalid or missing program code" });
+            return res.status(400).json({
+                success: false,
+                message: "Invalid or missing program code",
+            } as ProgramRequestData);
 
         const program = await programs.get(code);
         return res.status(program ? 200 : 404).json({
+            success: true,
             message: program
                 ? "Program retrieved successfully"
                 : "Unknown program with the provided program code",
             program,
-        });
+        } as ProgramRequestData);
     }
 );