+++ src/services/prisma.ts
@@ -5,6 +5,11 @@ import {
     StudentUser,
     FacultyUser,
     UserRole,
+    Program,
+    Course,
+    CourseSchedule,
+    CourseEnrollment,
+    Section,
 } from "../generated/prisma";
 
 const prisma = new PrismaClient();
@@ -12,10 +17,15 @@ export default prisma;
 
 export {
     Prisma,
+    UserRole,
 
     // Export models
     User as UserModel,
     StudentUser as StudentUserModel,
     FacultyUser as FacultyUserModel,
-    UserRole,
+    CourseSchedule as CourseScheduleModel,
+    CourseEnrollment as CourseEnrollmentModel,
+    Program as ProgramModel,
+    Course as CourseModel,
+    Section as SectionModel,
 };