+++ src/sis/course/@types.ts
@@ -1,0 +1,8 @@
+import { CourseModel } from "../../services/prisma";
+
+export interface CourseBasicRequestData {
+    message: string;
+}
+
+export type CourseRequestData = CourseBasicRequestData &
+    ({ success: true; course: CourseModel } | { success: false });