+++ src/sis/student/students.ts
@@ -1,6 +1,37 @@
+import { z } from "zod";
+
+import { GradeStatus } from "../../generated/prisma";
 import prisma from "../../services/prisma";
 
 class StudentManager {
+    protected _studentNumberSchema = z
+        .string()
+        .regex(/^[0-9]{4,}-[0-9]{5,}-[A-Z]{2,}-[0-9|X]{1,}/);
+    protected _enrollSchema = z.object({
+        studentNo: this._studentNumberSchema,
+        course: z.string(),
+        section: z.string().optional(),
+    });
+
+    protected _gradeSchema = z.object({
+        studentNo: this._studentNumberSchema,
+        course: z.string(),
+        grade: z.number(),
+        gradeStatus: z.enum(GradeStatus),
+    });
+
+    public studentNumberSchema() {
+        return this._studentNumberSchema;
+    }
+
+    public get enrollSchema() {
+        return this._enrollSchema;
+    }
+
+    public get gradeSchema() {
+        return this._gradeSchema;
+    }
+
     public async generateStudentNo() {
         return `${new Date().getFullYear()}-${(await prisma.studentUser.count()).toString().padStart(5, "0")}-${process.env.BRANCH_CODE!}-X`;
     }